refactor(event): 重构事件系统并重命名事件总线为事件模块
- 将 puniyu_bus 重命名为 puniyu_event 并更新所有相关引用
- 重构事件总线实现，使用 trait 定义事件总线接口
- 实现事件分发机制，支持处理器注册和事件匹配
- 更新依赖配置文件中的模块引用和组件名称
- 优化事件发送宏和消息事件构建器的实现
- 更新插件系统中的版本和作者信息处理逻辑

+++ crates/puniyu_types/src/handler.rs
@@ -14,7 +14,7 @@ pub trait Handler: Matcher {
 	fn name(&self) -> &str;
 
 	/// 优先级
-	fn rank(&self) -> u8 {
+	fn rank(&self) -> u32 {
 		5
 	}
